% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_table.R
\name{t_table}
\alias{t_table}
\title{t table function}
\usage{
t_table(alpha = 0.05, df = 1, tail = 1)
}
\arguments{
\item{numeric}{input alpha, degrees of freedom and either 1 or 2 tails}
}
\description{
This function returns t values for a given alpha level, degree of freedom, and one/two tails.
}
\examples{
t_table(alpha = .05, df = 3, tail = 1)
}
\keyword{t}
\keyword{t-test}
\keyword{table}
