% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_table.R
\name{z_table}
\alias{z_table}
\title{z table function}
\usage{
z_table(z = 0)
}
\arguments{
\item{numeric}{input a z-score}
}
\description{
This function returns body/tail proportions for a given z-score.
}
\examples{
z_table(z = 1.96)
}
\keyword{score}
\keyword{table}
\keyword{z}
